requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
requests==2.31.0
lxml==5.1.0
Pillow==10.2.0
